+++ src/migrations/m180520_120000_remove_duplicate_columns.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace putyourlightson\campaign\migrations;
+
+use craft\db\Migration;
+
+/**
+ * m180520_120000_remove_duplicate_columns migration.
+ */
+class m180520_120000_remove_duplicate_columns extends Migration
+{
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    public function safeUp()
+    {
+        $this->dropColumn('{{%campaign_contacts_campaigns}}', 'country');
+        $this->dropColumn('{{%campaign_contacts_campaigns}}', 'geoIp');
+        $this->dropColumn('{{%campaign_contacts_campaigns}}', 'device');
+        $this->dropColumn('{{%campaign_contacts_campaigns}}', 'os');
+        $this->dropColumn('{{%campaign_contacts_campaigns}}', 'client');
+
+        $this->dropColumn('{{%campaign_contacts_mailinglists}}', 'country');
+        $this->dropColumn('{{%campaign_contacts_mailinglists}}', 'geoIp');
+        $this->dropColumn('{{%campaign_contacts_mailinglists}}', 'device');
+        $this->dropColumn('{{%campaign_contacts_mailinglists}}', 'os');
+        $this->dropColumn('{{%campaign_contacts_mailinglists}}', 'client');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown(): bool
+    {
+        echo "m180520_120000_remove_duplicate_columns cannot be reverted.\n";
+
+        return false;
+    }
+}